requirements.txt
requests==2.32.3
psycopg2-binary==2.9.10
apache-airflow==2.10.2
python-dotenv==1.0.1
pydantic==2.9.2
